move queryfilter tests to their own file
